+++ test/unit/fuzz-tools.test.js
@@ -189,6 +189,7 @@ import * as liverV196 from '../../lib/liver-v196.js';
 import * as endoQuantV197 from '../../lib/endo-quant-v197.js';
 import * as subspecialtyV198 from '../../lib/subspecialty-v198.js';
 import * as myeloidPrognosisV199 from '../../lib/myeloid-prognosis-v199.js';
+import * as critcareSeverityV200 from '../../lib/critcare-severity-v200.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -322,6 +323,7 @@ const MODULES = {
   'endo-quant-v197.js': endoQuantV197,
   'subspecialty-v198.js': subspecialtyV198,
   'myeloid-prognosis-v199.js': myeloidPrognosisV199,
+  'critcare-severity-v200.js': critcareSeverityV200,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];